chore: organise explore page cards
