+++ resources/js/components/cards/data/explore/BrowseFolderCard.vue
