+++ resources/js/components/cards/data/explore/BrowseMediaCard.vue
